refactor: split layout-mode and multi-pane utils
`map-bottom-sheet-layout-preset.ts` no longer matched its name after
2f90a835 dropped the simple-mode preset. The remaining utilities split
cleanly into two concerns, so move them to dedicated files.

- src/utils/layout-mode.ts (new) — layout-mode threshold:
  `LayoutMode` type, `WIDE_VIEWPORT_MIN_WIDTH`, `isWideViewport`.
- src/utils/multi-pane.ts (new) — multi-pane internals:
  `MultiPaneOrientation` type, `MULTI_PANE_PANEL_SIZE`,
  `resolveMultiPaneOrientation`.
- `LayoutMode` moves from `use-layout-mode.ts` to `layout-mode.ts`
  (single source of truth). The hook re-exports it for convenience.
- Importers updated: `use-layout-mode.ts`, `use-multi-pane-orientation
  .ts`, `multi-pane-layout.tsx`.
- Tests split into `layout-mode.test.ts` and `multi-pane.test.ts`.
- Old `map-bottom-sheet-layout-preset.ts` and its test deleted (the
  deletions were already staged by `git rm`).
